Extract page navigation component into its own file
